\name{GeneralizedLambda-internal}
\alias{glCheckParsErrorMessage}
\alias{gl_fm5_distfunc}
\alias{gl_fmkl_distfunc}
\alias{gl_rs_distfunc}


\title{Internal Generalized Lambda Functions}
\description{
  Internal functions for the package GeneralizedLambda
}
\details{
  Functions which are either not intended to be called by the user or
  are waiting to be documented.
}
\keyword{ internal }
